api/requirements/base.txt
django>=1.10,<1.11
django-environ==0.4.1
psycopg2==2.6.2
djangorestframework==3.5.4

